requirements.txt
anyjson==0.3.3
celery==4.1.0
decorator==4.0.6
Django==1.10.4
django-ajax==0.2.0
django-celery==3.2.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-dajax==0.9.2
django-dajaxice==0.7
djangorestframework==3.6.4
flower==0.9.2
redis==2.10.5

